requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
paho-mqtt==2.1.0
SQLAlchemy==2.0.28
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn
eventlet